+++ App.tsx
@@ -1,52 +1,12 @@
-import {NavigationContainer} from '@react-navigation/native';
 import React, {useEffect} from 'react';
 import SplashScreen from 'react-native-splash-screen'; //import SplashScreen
-import {createNativeStackNavigator} from '@react-navigation/native-stack';
-import Home from './src/screens/Home';
-import Login from './src/screens/Login';
-import Register from './src/screens/Register';
-import {APP_NAME} from '@env';
-const Stack = createNativeStackNavigator();
+import Navigation from './src/components/Navigation';
 
 const App = () => {
   useEffect(() => {
     SplashScreen.hide(); //hides the splash screen on app load.
   }, []);
-  return (
-    <NavigationContainer>
-      <Stack.Navigator>
-        <Stack.Screen
-          name="Home"
-          component={Home}
-          options={{
-            title: process.env.REACT_APP_APP_NAME,
-            headerStyle: {
-              backgroundColor: '#237091',
-            },
-            headerTintColor: '#fff',
-            headerTitleStyle: {
-              fontWeight: 'bold',
-            },
-          }}
-        />
-        <Stack.Screen name="Login" component={Login} />
-        <Stack.Screen
-          name="Register"
-          component={Register}
-          options={{
-            title: process.env.REACT_APP_APP_NAME + ' - Register',
-            headerStyle: {
-              backgroundColor: '#237091',
-            },
-            headerTintColor: '#fff',
-            headerTitleStyle: {
-              fontWeight: 'bold',
-            },
-          }}
-        />
-      </Stack.Navigator>
-    </NavigationContainer>
-  );
+  return <Navigation />;
 };
 
 export default App;
