+++ package.json
@@ -11,6 +11,7 @@
     "build-dev": "dotenv -e .env react-app-rewired build"
   },
   "dependencies": {
+    "@react-native-async-storage/async-storage": "^1.18.1",
     "@react-navigation/native": "^6.1.6",
     "@react-navigation/native-stack": "^6.9.12",
     "formik": "^2.2.9",
