+++ src/components/Navigation.tsx
@@ -1,0 +1,47 @@
+import {NavigationContainer} from '@react-navigation/native';
+import React from 'react';
+import {createNativeStackNavigator} from '@react-navigation/native-stack';
+import Home from '../screens/Home';
+import Login from '../screens/Login';
+import Register from '../screens/Register';
+const Stack = createNativeStackNavigator();
+
+const Navigation = () => {
+  return (
+    <NavigationContainer>
+      <Stack.Navigator>
+        <Stack.Screen
+          name="Home"
+          component={Home}
+          options={{
+            title: process.env.REACT_APP_APP_NAME,
+            headerStyle: {
+              backgroundColor: '#237091',
+            },
+            headerTintColor: '#fff',
+            headerTitleStyle: {
+              fontWeight: 'bold',
+            },
+          }}
+        />
+        <Stack.Screen name="Login" component={Login} />
+        <Stack.Screen
+          name="Register"
+          component={Register}
+          options={{
+            title: process.env.REACT_APP_APP_NAME + ' - Register',
+            headerStyle: {
+              backgroundColor: '#237091',
+            },
+            headerTintColor: '#fff',
+            headerTitleStyle: {
+              fontWeight: 'bold',
+            },
+          }}
+        />
+      </Stack.Navigator>
+    </NavigationContainer>
+  );
+};
+
+export default Navigation;
